refactor: Extract the query caches into a _QueryCacheStore in their own file
